api separated from config
