+++ components/plots/zoomedPlots/zoomedOverlayPlots/utils.ts
@@ -1,6 +1,6 @@
 import { ParamsForApiProps } from '../../../../containers/display/interfaces';
 
-import { get_plot_with_overlay, get_plot_url } from '../../../../config/config';
+import { get_plot_with_overlay, get_plot_url } from '../../../../api/oldApi';
 
 export const get_plot_source = (params_for_api: ParamsForApiProps) => {
   if (params_for_api.overlay === 'onSide') {
