+++ components/plots/zoomedPlots/zoomedOverlayPlots/zoomedOverlaidJSROOTPlot.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import cleanDeep from 'clean-deep';
 import { useRouter } from 'next/router';
 
-import { get_jroot_plot, functions_config } from '../../../../config/config';
+import { functions_config } from '../../../../config/config';
+import { get_jroot_plot } from '../../../../api/oldApi';
 import {
   ParamsForApiProps,
   TripleProps,
