+++ plotsLocalOverlayPageComponents/plots/jsroot/singleJSROOTPlot.tsx
@@ -1,6 +1,6 @@
 import { Tooltip } from 'antd';
 import * as React from 'react'
-import { get_jroot_plot } from '../../../config/config';
+import { get_jroot_plot } from '../../../api/oldApi';
 
 import { ImageDiv } from '../../../containers/display/styledComponents'
 import { useRequest } from '../../../hooks/useRequest';
